+++ src/web/twig/Environment.php
@@ -34,26 +34,6 @@ class Environment extends TwigEnvironment
     /**
      * @inheritdoc
      */
-    public function loadTemplate($name, $index = null)
-    {
-        try {
-            /** @noinspection PhpInternalEntityUsedInspection */
-            return parent::loadTemplate($name, $index);
-        } catch (Error $e) {
-            if (Craft::$app->getConfig()->getGeneral()->suppressTemplateErrors) {
-                // Just log it and return an empty template
-                Craft::$app->getErrorHandler()->logException($e);
-
-                return Craft::$app->getView()->renderString('');
-            }
-
-            throw $e;
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
     public function compileSource(Source $source)
     {
         Craft::beginProfile($source->getName(), __METHOD__);